Revert ViewerCollector and add twitch CAP negotiation for membership and commands
membership CAP will enable JOIN, PART, MODE and NAMES events in the bot again.
Twitch made a change which caused these events to not being sent by default
anymore.
commands CAP gives us events in case of HOSTMODE and so on.

Reverted the viewer collector to work with JOIN and PART Events again.