Rework raffle and introduce concurrent save handling
Now concurrent saves of viewers are handeled correctly.
A winners session points are reset on a win. This save operation is
FORCED.
If we didn't force that save, we could run into a situation where
the winner's points get reset and after the reset save, the winner
gets points awareded in the points collector. The points collector
could have the old wallet values and thus award a huge amount of
points to the winner, which whould ultimately restore the high
winning chance that viewer had before instead of keeping it reset
after a win.