THIS COMMIT INTRODUCES JAVA 8, rework raffle command
Reworked raffle command and added some tests. Raffles are using
Java8 streams for filtering the participants in parallel.

Fixed a bug in the raffle where participants with 0 sessionPoints
could not only win, but even had a way higher chance of winning
than other participants if the participant pool was small.